requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
altair==4.0.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chromadb==0.3.21
click==8.1.7
clickhouse-connect==0.7.7
colorama==0.4.6
dataclasses-json==0.5.14
duckdb==0.10.1
entrypoints==0.4
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
hnswlib==0.8.0
httptools==0.6.1
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.0.184
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.6
numpy==1.24.4
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.0.3
Pillow==9.5.0
pip==22.0.4
pkgutil_resolve_name==1.3.10
posthog==3.5.0
protobuf==4.25.3
pyarrow==15.0.2
pydantic==1.10.15
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.6.1
setuptools==58.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
streamlit==1.24.0
streamlit-chat-animated==0.0.3.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.4.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.1
uvicorn==0.29.0
validators==0.28.0
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
